Split out FrontendInputsAndOutputs
